fix(presupuestos): corrige dirección perdida y cantidad en unidad inventada
- POST /presupuesto/nuevo insertaba cliente_direccion vía json_decode() de un
  campo que los callers (presupuesto.vue, 19print_app) mandan como texto
  plano -- json_decode() sobre texto no-JSON devuelve null en silencio y la
  dirección quedaba siempre vacía. Se usa el valor crudo del form, igual
  que cliente_nombre/observaciones/fecha_entrega.
- presupuestos_productos.cantidad era INTEGER: 19print_app facturaba en una
  unidad inventada ("decímetros", cantidad x10 / precio /10) para esquivar
  la restricción, invisible para el resto de Ninesys (el catálogo etiqueta
  sus tramos de precio en metros). Columna ampliada a DECIMAL(6,1), misma
  precisión que ordenes_productos.cantidad -- actualizado en ambas
  plantillas de empresa nueva (MySQL y Postgres).

+++ public/model/create_new_company_api_emp_N_postgres.sql
@@ -1142,7 +1142,7 @@ CREATE TABLE presupuestos_productos (
   id_category INTEGER DEFAULT NULL ,
   category_name varchar(20) DEFAULT NULL ,
   name varchar(240) DEFAULT NULL ,
-  cantidad INTEGER NOT NULL DEFAULT 0 ,
+  cantidad DECIMAL(6,1) NOT NULL DEFAULT 0 ,
   talla varchar(32) DEFAULT NULL ,
   corte varchar(32) DEFAULT NULL ,
   tela varchar(128) DEFAULT NULL ,
